Move Register Views tests to register package
